Moved TreePrettyPrinter to tree package
The old import still works, but throws a warning. Perfect!
