+++ nltk/tree/__init__.py
@@ -25,5 +25,6 @@ from nltk.tree.immutable import (
 )
 from nltk.tree.parented import MultiParentedTree, ParentedTree
 from nltk.tree.parsing import bracket_parse, sinica_parse
+from nltk.tree.prettyprinter import TreePrettyPrinter
 from nltk.tree.probabilistic import ProbabilisticTree
 from nltk.tree.tree import Tree
